Register handlebar helper in form-utils
