+++ components/org.wso2.carbon.siddhi.editor.core/src/main/resources/web/js/design-view/form-builder/forms/table-form.js
@@ -41,37 +41,6 @@ define(['log', 'jquery', 'lodash', 'attribute', 'table', 'storeAnnotation', 'des
         const defaultStoreType = "in-memory";
         const rdbmsStoreType = "rdbms";
 
-        /** Handlebar helper to render heading for the form */
-        Handlebars.registerHelper('addTitle', function (id) {
-            return id.charAt(0).toUpperCase() + id.slice(1);
-        });
-
-        /** Handlebar helper to compare if the id is "source" or "sink" or "store" */
-        Handlebars.registerHelper('ifSourceOrSinkOrStore', function (id, div) {
-            if (id === "source" || id === "sink" || id === "store") {
-                return div.fn(this);
-            }
-            return div.inverse(this);
-        });
-
-        /** Generates the current index of the option being rendered */
-        Handlebars.registerHelper('sum', function () {
-            return Array.prototype.slice.call(arguments, 0, -1).reduce((acc, num) => acc += num);
-        });
-
-        /** Handlebar helper to check if the index is equivalent to half the length of the option's array */
-        Handlebars.registerHelper('isDivisor', function (index, options) {
-            var divLength = Math.ceil(options.length / 2);
-            return index === divLength;
-        });
-
-        /** Handlebar helper to check id is equivalent to a given string */
-        Handlebars.registerHelper('ifId', function (id, name, div) {
-            if (id === name) {
-                return div.fn(this);
-            }
-            return div.inverse(this);
-        });
 
         /** Function to manage the attribute navigations */
         var changeAtrributeNavigation = function () {
